requirements.txt
playwright
beautifulsoup4
requests
pyserial-asyncio
